requirements.txt
torch
transformers
sentence-transformers
pypdf
faiss-cpu
fastapi
uvicorn
langchain
pydantic
python-dotenv
tqdm
numpy
pandas
scikit-learn
streamlit